+++ src/id_translation/dio/_sequence.py
@@ -5,7 +5,7 @@ import pandas as pd
 from rics.collections.misc import as_list
 
 from ..offline import TranslationMap
-from ..ttypes import IdType, NameType, SourceType
+from ..types import IdType, NameType, SourceType
 from . import DataStructureIO
 from .exceptions import NotInplaceTranslatableError